style(launch): move launch errors closer to the code (#4995)
* move docker and launch tests out of main errors

* lint
