+++ tests/pytest_tests/unit_tests_old/tests_launch/test_launch_aws.py
@@ -17,6 +17,7 @@ from wandb.sdk.launch.runner.aws import (
     get_region,
     validate_sagemaker_requirements,
 )
+from wandb.sdk.launch.utils import LaunchError
 
 from tests.pytest_tests.unit_tests_old.utils import fixture_open
 
@@ -232,7 +233,7 @@ def test_launch_aws_sagemaker_launch_fail(
     kwargs["uri"] = uri
     kwargs["api"] = api
 
-    with pytest.raises(wandb.errors.LaunchError) as e_info:
+    with pytest.raises(LaunchError) as e_info:
         launch.run(**kwargs)
     assert "Unable to create training job" in str(e_info.value)
 
@@ -261,7 +262,7 @@ def test_launch_aws_sagemaker_push_image_fail_none(
     kwargs["uri"] = uri
     kwargs["api"] = api
 
-    with pytest.raises(wandb.errors.LaunchError) as e_info:
+    with pytest.raises(LaunchError) as e_info:
         launch.run(**kwargs)
     assert "Failed to push image to repository" in str(e_info.value)
 
@@ -291,7 +292,7 @@ def test_launch_aws_sagemaker_push_image_fail_err_msg(
     kwargs["uri"] = uri
     kwargs["api"] = api
 
-    with pytest.raises(wandb.errors.LaunchError) as e_info:
+    with pytest.raises(LaunchError) as e_info:
         launch.run(**kwargs)
     assert "I regret to inform you, that I have failed" in str(e_info.value)
 
@@ -433,7 +434,7 @@ def test_failed_aws_cred_login(
         kwargs["uri"] = uri
         kwargs["api"] = api
 
-        with pytest.raises(wandb.errors.LaunchError):
+        with pytest.raises(LaunchError):
             launch.run(**kwargs)
 
 
@@ -492,7 +493,7 @@ def test_aws_get_region_file_fail_no_section(runner, monkeypatch):
             cuda=None,
             run_id=None,
         )
-        with pytest.raises(wandb.errors.LaunchError) as e_info:
+        with pytest.raises(LaunchError) as e_info:
             get_region(launch_project.resource_args)
         assert (
             str(e_info.value)
@@ -521,7 +522,7 @@ def test_aws_get_region_file_fail_no_file(runner, monkeypatch):
             cuda=None,
             run_id=None,
         )
-        with pytest.raises(wandb.errors.LaunchError) as e_info:
+        with pytest.raises(LaunchError) as e_info:
             get_region(launch_project.resource_args)
         assert (
             str(e_info.value)
@@ -550,7 +551,7 @@ def test_no_sagemaker_resource_args(
         kwargs["uri"] = uri
         kwargs["api"] = api
         kwargs["resource_args"].pop("sagemaker", None)
-        with pytest.raises(wandb.errors.LaunchError) as e_info:
+        with pytest.raises(LaunchError) as e_info:
             launch.run(**kwargs)
         assert (
             str(e_info.value)
@@ -582,7 +583,7 @@ def test_no_OuputDataConfig(
         kwargs["uri"] = uri
         kwargs["api"] = api
         kwargs["resource_args"]["sagemaker"].pop("OutputDataConfig", None)
-        with pytest.raises(wandb.errors.LaunchError) as e_info:
+        with pytest.raises(LaunchError) as e_info:
             launch.run(**kwargs)
         assert (
             str(e_info.value)
@@ -612,7 +613,7 @@ def test_no_StoppingCondition(
         kwargs["api"] = api
         kwargs["resource_args"]["sagemaker"].pop("StoppingCondition", None)
 
-        with pytest.raises(wandb.errors.LaunchError) as e_info:
+        with pytest.raises(LaunchError) as e_info:
             launch.run(**kwargs)
         assert (
             str(e_info.value)
@@ -642,7 +643,7 @@ def test_no_ResourceConfig(
         kwargs["api"] = api
         kwargs["resource_args"]["sagemaker"].pop("ResourceConfig", None)
 
-        with pytest.raises(wandb.errors.LaunchError) as e_info:
+        with pytest.raises(LaunchError) as e_info:
             launch.run(**kwargs)
         assert (
             str(e_info.value)
@@ -672,7 +673,7 @@ def test_no_RoleARN(
         kwargs["api"] = api
         kwargs["resource_args"]["sagemaker"].pop("RoleArn", None)
 
-        with pytest.raises(wandb.errors.LaunchError) as e_info:
+        with pytest.raises(LaunchError) as e_info:
             launch.run(**kwargs)
         assert (
             str(e_info.value)
@@ -684,11 +685,11 @@ def test_no_RoleARN(
 def test_validate_sagemaker_requirements():
     given_sagemaker_args = {}
     registry_config = {}
-    with pytest.raises(wandb.errors.LaunchError):
+    with pytest.raises(LaunchError):
         validate_sagemaker_requirements(given_sagemaker_args, registry_config)
 
     registry_config["ecr-repo-provider"] = "gcp"
-    with pytest.raises(wandb.errors.LaunchError):
+    with pytest.raises(LaunchError):
         validate_sagemaker_requirements(given_sagemaker_args, registry_config)
 
 
@@ -699,11 +700,11 @@ def test_get_ecr_repository_url():
     }
     given_sagemaker_args = {}
     registry_config = {}
-    with pytest.raises(wandb.errors.LaunchError):
+    with pytest.raises(LaunchError):
         get_ecr_repository_url(client, given_sagemaker_args, registry_config)
 
     given_sagemaker_args["EcrRepoName"] = {"asd": 123}
-    with pytest.raises(wandb.errors.LaunchError):
+    with pytest.raises(LaunchError):
         get_ecr_repository_url(client, given_sagemaker_args, registry_config)
 
     given_sagemaker_args["EcrRepoName"] = "test_repo_name"
