+++ tests/pytest_tests/unit_tests_old/tests_launch/test_launch_jobs.py
@@ -7,9 +7,10 @@ import pytest
 import wandb
 import wandb.sdk.launch._project_spec as _project_spec
 import wandb.sdk.launch.launch as launch
-from wandb.errors import CommError, LaunchError
+from wandb.errors import CommError
 from wandb.sdk.data_types._dtypes import TypeRegistry
 from wandb.sdk.launch.launch_add import launch_add
+from wandb.sdk.launch.utils import LaunchError
 
 from tests.pytest_tests.unit_tests_old import utils
