+++ tests/pytest_tests/unit_tests_old/tests_launch/test_launch_utils.py
@@ -2,8 +2,7 @@ import random
 from typing import List
 
 import pytest
-from wandb.errors import LaunchError
-from wandb.sdk.launch.utils import diff_pip_requirements, parse_wandb_uri
+from wandb.sdk.launch.utils import LaunchError, diff_pip_requirements, parse_wandb_uri
 
 REQUIREMENT_FILE_BASIC: List[str] = [
     "package-one==1.0.0",
