+++ wandb/errors/__init__.py
@@ -3,13 +3,9 @@ __all__ = [
     "UsageError",
     "CommError",
     "LogError",
-    "DockerError",
     "LogMultiprocessError",
     "MultiprocessError",
     "RequireError",
-    "ExecutionError",
-    "LaunchError",
-    "SweepError",
     "WaitTimeoutError",
     "ContextCancelledError",
 ]
@@ -67,56 +63,6 @@ class MultiprocessError(Error):
 
 class RequireError(Error):
     """Raised when wandb.require() fails"""
-
-    pass
-
-
-class ExecutionError(Error):
-    """Generic execution exception"""
-
-    pass
-
-
-class DockerError(Error):
-    """Raised when attempting to execute a docker command"""
-
-    def __init__(
-        self,
-        command_launched: List[str],
-        return_code: int,
-        stdout: Optional[bytes] = None,
-        stderr: Optional[bytes] = None,
-    ) -> None:
-        command_launched_str = " ".join(command_launched)
-        error_msg = (
-            f"The docker command executed was `{command_launched_str}`.\n"
-            f"It returned with code {return_code}\n"
-        )
-        if stdout is not None:
-            error_msg += f"The content of stdout is '{stdout.decode()}'\n"
-        else:
-            error_msg += (
-                "The content of stdout can be found above the "
-                "stacktrace (it wasn't captured).\n"
-            )
-        if stderr is not None:
-            error_msg += f"The content of stderr is '{stderr.decode()}'\n"
-        else:
-            error_msg += (
-                "The content of stderr can be found above the "
-                "stacktrace (it wasn't captured)."
-            )
-        super().__init__(error_msg)
-
-
-class LaunchError(Error):
-    """Raised when a known error occurs in wandb launch"""
-
-    pass
-
-
-class SweepError(Error):
-    """Raised when a known error occurs with wandb sweeps"""
 
     pass
