+++ wandb/sdk/launch/builder/kaniko.py
@@ -10,7 +10,6 @@ import kubernetes  # type: ignore
 from kubernetes import client
 
 import wandb
-from wandb.errors import LaunchError
 from wandb.sdk.launch.builder.abstract import AbstractBuilder
 from wandb.util import get_module
 
@@ -20,7 +19,12 @@ from .._project_spec import (
     create_metadata_file,
     get_entry_point_command,
 )
-from ..utils import LOG_PREFIX, get_kube_context_and_api_client, sanitize_wandb_api_key
+from ..utils import (
+    LOG_PREFIX,
+    LaunchError,
+    get_kube_context_and_api_client,
+    sanitize_wandb_api_key,
+)
 from .build import _create_docker_build_ctx, generate_dockerfile
 
 _DEFAULT_BUILD_TIMEOUT_SECS = 1800  # 30 minute build timeout
