+++ wandb/sdk/launch/builder/loader.py
@@ -1,7 +1,7 @@
 import logging
 from typing import Any, Dict, List
 
-from wandb.errors import LaunchError
+from wandb.sdk.launch.utils import LaunchError
 
 from .abstract import AbstractBuilder
