+++ wandb/sdk/launch/launch.py
@@ -6,7 +6,6 @@ import yaml
 
 import wandb
 from wandb.apis.internal import Api
-from wandb.errors import ExecutionError, LaunchError
 
 from ._project_spec import create_project_from_spec, fetch_and_validate_project
 from .agent import LaunchAgent
@@ -18,6 +17,8 @@ from .utils import (
     LAUNCH_CONFIG_FILE,
     LAUNCH_DEFAULT_PROJECT,
     PROJECT_SYNCHRONOUS,
+    ExecutionError,
+    LaunchError,
     construct_launch_spec,
     validate_launch_spec_source,
 )
