+++ wandb/sdk/launch/launch_add.py
@@ -4,7 +4,6 @@ from typing import Any, Dict, List, Optional
 import wandb
 import wandb.apis.public as public
 from wandb.apis.internal import Api
-from wandb.errors import LaunchError
 from wandb.sdk.launch._project_spec import (
     compute_command_args,
     create_project_from_spec,
@@ -13,6 +12,7 @@ from wandb.sdk.launch.builder.build import build_image_from_project
 from wandb.sdk.launch.utils import (
     LAUNCH_DEFAULT_PROJECT,
     LOG_PREFIX,
+    LaunchError,
     construct_launch_spec,
     validate_launch_spec_source,
 )
