+++ wandb/sdk/launch/runner/gcp_vertex.py
@@ -11,13 +11,12 @@ if False:
 import yaml
 
 import wandb
-from wandb.errors import LaunchError
 from wandb.util import get_module
 
 from .._project_spec import LaunchProject, get_entry_point_command
 from ..builder.abstract import AbstractBuilder
 from ..builder.build import construct_gcp_registry_uri, get_env_vars_dict
-from ..utils import LOG_PREFIX, PROJECT_SYNCHRONOUS, run_shell
+from ..utils import LOG_PREFIX, PROJECT_SYNCHRONOUS, LaunchError, run_shell
 from .abstract import AbstractRun, AbstractRunner, Status
 
 GCP_CONSOLE_URI = "https://console.cloud.google.com"
