+++ wandb/sdk/launch/runner/local_process.py
@@ -3,13 +3,13 @@ import shlex
 from typing import Any, List, Optional
 
 import wandb
-from wandb.errors import LaunchError
 
 from .._project_spec import LaunchProject, get_entry_point_command
 from ..builder.build import get_env_vars_dict
 from ..utils import (
     LOG_PREFIX,
     PROJECT_SYNCHRONOUS,
+    LaunchError,
     _is_wandb_uri,
     download_wandb_python_deps,
     parse_wandb_uri,
